style(#585): move api-structs
Moded structs of the api-endpoints into a new
rust-library.
