+++ src/binaries/hanami/src/api/http_endpoints/auth/mod.rs
@@ -14,4 +14,3 @@
 
 pub mod create_token_v1_0;
 pub mod renew_token_v1_0;
-mod auth_structs;
