+++ src/binaries/hanami/src/api/http_endpoints/cluster/create_cluster_v1_0.rs
@@ -23,8 +23,7 @@ use crate::database::cluster_table;
 use crate::core::cluster_handler;
 
 use hanami_common::error::HanamiError;
-
-use super::cluster_structs::{ClusterCreateReq, ClusterResp};
+use hanami_structs::cluster_structs::{ClusterCreateReq, ClusterResp};
 
 #[api_operation(
     tag = "cluster",
