+++ src/binaries/hanami/src/api/http_endpoints/cluster/list_cluster_v1_0.rs
@@ -20,7 +20,7 @@ use crate::api::errors::ErrorResponse;
 use crate::api::user_context::UserContext;
 use crate::database::cluster_table;
 
-use super::cluster_structs::{ClusterBasicResp, ClusterListResp};
+use hanami_structs::cluster_structs::{ClusterBasicResp, ClusterListResp};
 
 #[api_operation(
     tag = "cluster",
