+++ src/binaries/hanami/src/api/http_endpoints/cluster/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod cluster_structs;
 pub mod create_cluster_v1_0;
 pub mod get_cluster_v1_0;
 pub mod list_cluster_v1_0;
