+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/list_task_v1_0.rs
@@ -23,8 +23,7 @@ use crate::database::task_table;
 use crate::database::cluster_table;
 
 use hanami_common::enums;
-
-use super::task_structs::{TaskBasicResp, TaskListResp, TaskType, TaskState};
+use hanami_structs::task_structs::{TaskBasicResp, TaskListResp, TaskType, TaskState};
 
 #[api_operation(
     tag = "task",
