+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/mod.rs
@@ -16,7 +16,6 @@ pub mod create_train_task_v1_0;
 pub mod create_request_task_v1_0;
 pub mod get_task_v1_0;
 pub mod list_task_v1_0;
-pub mod task_structs;
 pub mod checkpoint_save_task_v1_0;
 pub mod checkpoint_restore_task_v1_0;
 pub mod abort_task_v1_0;
