+++ src/binaries/hanami/src/api/http_endpoints/dataset/get_dataset_v1_0.rs
@@ -24,8 +24,7 @@ use crate::database::dataset_table;
 
 use hanami_common::enums;
 use hanami_dataset::dataset_io::read_data_set_file;
-
-use super::dataset_structs::DatasetResp;
+use hanami_structs::dataset_structs::DatasetResp;
 
 #[api_operation(
     tag = "dataset",
