+++ src/binaries/hanami/src/api/http_endpoints/dataset/mod.rs
@@ -16,5 +16,4 @@ pub mod create_dataset_v1_0;
 pub mod get_dataset_v1_0;
 pub mod list_dataset_v1_0;
 pub mod delete_dataset_v1_0;
-pub mod dataset_structs;
 pub mod check_dataset_v1_0;
