+++ src/binaries/hanami/src/api/http_endpoints/project/create_project_v1_0.rs
@@ -19,9 +19,9 @@ use apistos::api_operation;
 use crate::api::user_context::UserContext;
 use crate::api::errors::ErrorResponse;
 use crate::database::project_table;
-use hanami_common::enums;
 
-use super::project_structs::{ProjectCreateReq, ProjectResp};
+use hanami_common::enums;
+use hanami_structs::project_structs::{ProjectCreateReq, ProjectResp};
 
 #[api_operation(
     tag = "project",
