+++ src/binaries/hanami/src/api/http_endpoints/project/delete_project_v1_0.rs
@@ -19,6 +19,7 @@ use actix_web::web::Path;
 use crate::api::errors::ErrorResponse;
 use crate::api::user_context::UserContext;
 use crate::database::project_table;
+
 use hanami_common::enums;
 
 #[api_operation(
