+++ src/binaries/hanami/src/api/http_endpoints/project/mod.rs
@@ -16,4 +16,3 @@ pub mod create_project_v1_0;
 pub mod list_project_v1_0;
 pub mod get_project_v1_0;
 pub mod delete_project_v1_0;
-mod project_structs;
