+++ src/binaries/hanami/src/api/http_endpoints/user/get_user_v1_0.rs
@@ -19,9 +19,9 @@ use apistos::api_operation;
 use crate::api::errors::ErrorResponse;
 use crate::api::user_context::UserContext;
 use crate::database::user_table;
-use hanami_common::enums;
 
-use super::user_structs::UserResp;
+use hanami_common::enums;
+use hanami_structs::user_structs::UserResp;
 
 #[api_operation(
     tag = "user",
