+++ src/libs/rust/hanami_structs/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "hanami_structs"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+# openapi-generator
+schemars = { package = "apistos-schemars", version = "0.8", features = ["chrono", "uuid1", "url", "rust_decimal"] }
+apistos = { version = "0.5", features = ["extras"] }
+# json
+serde = { version = "1", features = ["derive"] }
+# common stuff
+uuid = { version = "1.17.0", features = ["v4"] }
