+++ src/libs/rust/hanami_structs/src/auth_structs.rs
