+++ src/libs/rust/hanami_structs/src/lib.rs
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod get_checkpoint_v1_0;
-pub mod list_checkpoint_v1_0;
-pub mod delete_checkpoint_v1_0;
+pub mod auth_structs;
+pub mod checkpoint_structs;
+pub mod cluster_structs;
+pub mod dataset_structs;
+pub mod project_structs;
+pub mod task_structs;
+pub mod user_structs;
