+++ src/libs/rust/hanami_structs/src/user_structs.rs
